Adding a DataSet
Creates the DataSet add page, and implements file uploading from the
browser.  FileReader JS API is accessed through ports.  There is a lib
that uses Native code to read from this API, but, I decided to go with
the ports implementation in case the upcoming .19 release breask things.
https://github.com/simonh1000/file-reader

The tab control is not pulled out as reusable, since we only appear to
be using a tab interface in one place.