Rewrite column metadata editor to allow for per row editing.
The editor now needs a save command, to be supplied by the container,
 which will end up making the remote http call.
This lets us show the error message within the editor if something goes wrong.

Tracking what changes are being made is a bit complex since we need to ensure that
there is only one target at a given time, accoss three different dictionaries of columns.